Move fetchAllSavedAlbums to spotifyAPI.js
